pyproject.toml
[project]
name = "tweet-scraper"
version = "0.2.0"
description = "Fetch Twitter/X tweets and export to Markdown"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "vcrpy>=5.1.0",
]

[project.scripts]
tweet-scraper = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
